shift client and resolver from 1.0 to v0 directory
